requirements.txt
flask
pymongo[srv]==3.11.0

python-dotenv